+++ arbitragelab/copula_approach/__init__.py
@@ -2,8 +2,12 @@
 This module implements Copula-based Statistical Arbitrage strategies.
 """
 
-from arbitragelab.copula_approach.copula_calculation import (find_marginal_cdf, ml_theta_hat, log_ml,
-                                                             sic, aic, hqic)
-from arbitragelab.copula_approach.copula_generate import (Copula, Gumbel, Frank, Clayton, Joe, N13,
-                                                          N14, Gaussian, Student, Switcher)
+from arbitragelab.copula_approach.copula_calculation import (
+    find_marginal_cdf, ml_theta_hat, log_ml, sic, aic, hqic, construct_ecdf_lin, fit_nu_for_t_copula, scad_penalty,
+    scad_derivative, adjust_weights)
+from arbitragelab.copula_approach.copula_generate import (
+    Copula, Gumbel, Frank, Clayton, Joe, N13, N14, Gaussian, Student, Switcher)
 from arbitragelab.copula_approach.copula_strategy import CopulaStrategy
+from arbitragelab.copula_approach.copula_strategy_basic import BasicCopulaStrategy
+from arbitragelab.copula_approach.copula_strategy_mpi import CopulaStrategyMPI
+from arbitragelab.copula_approach.copula_generate_mixedcopula import (MixedCopula, CFGMixCop, CTGMixCop)