+++ docs/source/changelog.rst
@@ -12,11 +12,18 @@ Changelog
 ..
     For Help: https://releases.readthedocs.io/en/latest/index.html
 
+* :feature:`28` CopulaStrategy replaced with improved BasicCopulaStrategy in the Copula Approach Module.
+* :feature:`28` Support of Clayton-Frank-Gumbel and Clayton-Student-Gumbel mixed copulas added to the Copula Approach Module.
+* :feature:`28` Mispricing Index Trading Strategy added to the Copula Approach Module.
+* :feature:`28` Quick Pairs Selection and ECDF added to the Copula Approach Module.
+* :support:`28` Updated Copula Brief Intro and added Copula Deeper Intro to documentation.
+* :support:`28` Mispricing Index Trading Strategy, Quick Pairs Selection and ECDF documentation.
+* :support:`28` Equity Curve Convention documentation.
+* :feature:`26` Multivariate Cointegration strategy (Galenko et al. 2010) added to the Cointegration Approach Module.
+* :support:`26` Multivariate Cointegration strategy documentation.
 * :support:`35` Upgrade ArbitrageLab requirements versions (numpy==1.20.1, matplotlib==3.3.4
   pandas==1.1.5, scikit-learn==0.24.1, scipy==1.6.0, statsmodels==0.12.2).
 * :bug:`34` Data Importer, Distance Approach, ML Approach modules imports were not exposed.
-* :feature:`26` Multivariate Cointegration strategy (Galenko et al. 2010) added to the Cointegration Approach Module.
-* :support:`26` Multivariate Cointegration strategy documentation.
 
 * :release:`0.2.2 <2020-12-24>`
 * :bug:`32` Copulas module imports were not exposed.